refactor:重构查询、清理、交接 # Reviewed, transaction id: 75361
