+++ saas/backend/api/external/staff_movement/urls.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 """
-TencentBlueKing is pleased to support the open source community by making 蓝鲸智云-权限中心(BlueKing-IAM) available.
+TencentBlueKing is pleased to support the open source community by making 蓝鲸智云 - 权限中心 (BlueKing-IAM) available.
 Copyright (C) 2017-2021 THL A29 Limited, a Tencent company. All rights reserved.
 Licensed under the MIT License (the "License"); you may not use this file except in compliance with the License.
 You may obtain a copy of the License at http://opensource.org/licenses/MIT
@@ -8,9 +8,13 @@ Unless required by applicable law or agreed to in writing, software distributed 
 an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 """
-from django.apps import AppConfig
 
+from django.urls import path
 
-class ResignAPIConfig(AppConfig):
-    name = "backend.api.resign"
-    label = "api_resign"
+from backend.api.external.staff_movement import views
+
+urlpatterns = [
+    path("getAssets", views.GetAssetsViewSet.as_view({"post": "list"}), name="resign_list"),
+    path("handover", views.ResignHandoverViewSet.as_view({"post": "handover"}), name="handover"),
+    path("recycle", views.RecycleViewSet.as_view({"post": "recycle"}), name="recycle"),
+]
