+++ saas/urls.py
@@ -34,7 +34,6 @@ schema_view = get_schema_view(
 )
 
 urlpatterns = [
-    url(r"^api/resign/", include("backend.api.resign.urls")),
     # backend apps url
     url(
         r"^api/v1/",
@@ -65,6 +64,8 @@ urlpatterns = [
                 url(r"^iam/", include("backend.iam.urls")),
                 # notice
                 url(r"^{}".format(config.ENTRANCE_URL), include(("bk_notice_sdk.urls", "notice"), namespace="notice")),
+                # 为外部系统提供API
+                url(r"^external/", include("backend.api.external.urls")),
             ]
         ),
     ),
